# [gwd] Static caching for database