--- a/configure.ml
+++ b/configure.ml
@@ -8,11 +8,14 @@ let rm = ref ""
 let ext = ref ""
 let os_type = ref ""
 let installed pkg = 0 = Sys.command ("ocamlfind query -qo -qe " ^ pkg)
+let nnp_compiler = 1 = Sys.command "$(ocamlc -config-var naked_pointers)"
 let errmsg = "usage: " ^ Sys.argv.(0) ^ " [options]"
 let api = ref false
 let sosa = ref `None
 let gwdb = ref `None
 let syslog = ref false
+let caching = ref false
+let set_caching () = caching := true
 let set_api () = api := true
 let set_syslog () = syslog := true
 
@@ -56,6 +59,9 @@ let speclist =
       Arg.Unit set_sosa_zarith,
       " Use Sosa module implementation based on `zarith` library" );
     ("--syslog", Arg.Unit set_syslog, " Log gwd errors using syslog");
+    ( "--gwd-caching",
+      Arg.Unit set_caching,
+      " Enable database preloading for gwd" );
   ]
   |> List.sort compare |> Arg.align
 
@@ -102,6 +108,23 @@ let () =
         (os_type, " -D UNIX", "", "/bin/rm -f", "strip")
     | _ -> ("Win", " -D WINDOWS", ".exe", "rm -f", "true")
   in
+  let ancient_lib, ancient_file =
+    let no_cache = ("", "gw_ancient.dum.ml") in
+    if nnp_compiler then
+      if installed "ancient" then ("ancient", "gw_ancient.wrapped.ml")
+      else (
+        if !caching then
+          Printf.eprintf
+            "Warning: ocaml-ancient not installed. Cannot enable database \
+             caching.\n";
+        no_cache)
+    else (
+      if !caching then
+        Printf.eprintf
+          "Warning: Compiler not set to no-naked-pointers. Cannot enable \
+           database caching.\n";
+      no_cache)
+  in
   let ch = open_out "Makefile.config" in
   let writeln s = output_string ch @@ s ^ "\n" in
   let var name value = writeln @@ name ^ "=" ^ value in
@@ -119,4 +142,6 @@ let () =
   var "SYSLOG_PKG" syslog_pkg;
   var "DUNE_DIRS_EXCLUDE" !dune_dirs_exclude;
   var "DUNE_PROFILE" dune_profile;
+  var "ANCIENT_LIB" ancient_lib;
+  var "ANCIENT_FILE" ancient_file;
   close_out ch