--- a/geneweb.opam
+++ b/geneweb.opam
@@ -24,6 +24,7 @@ depends: [
   "markup"
   "num"
   "ocaml" {>= "4.08"}
+  "ocaml-option-nnp"
   "ocamlformat" {dev & = "0.24.1"}
   "ounit"
   "ppx_blob"