--- a/geneweb.opam
+++ b/geneweb.opam
@@ -13,7 +13,6 @@ bug-reports: "https://github.com/geneweb/geneweb/issues"
 depends: [
   "dune" {>= "2.9"}
   "alcotest" {with-test}
-  "ancient"
   "benchmark"
   "calendars" {= "1.0.0"}
   "camlp5" {>= "8.00.01"}
@@ -24,7 +23,6 @@ depends: [
   "markup"
   "num"
   "ocaml" {>= "4.08"}
-  "ocaml-option-nnp"
   "ocamlformat" {dev & = "0.24.1"}
   "ounit"
   "ppx_blob"
@@ -40,6 +38,7 @@ depends: [
   "zarith"
   "odoc" {with-doc}
 ]
+depopts: ["ocaml-option-nnp" "ancient"]
 dev-repo: "git+https://github.com/geneweb/geneweb.git"
 build: [
   [ "ocaml" "./configure.ml" "--release" ]