--- a/lib/ancient/gw_ancient.dum.ml
+++ b/lib/ancient/gw_ancient.dum.ml
@@ -1,0 +1,7 @@
+let is_available = false
+
+type _ ancient = unit
+
+let mark _ = assert false
+let follow _ = assert false
+let delete _ = assert false