--- a/lib/ancient/gw_ancient.mli
+++ b/lib/ancient/gw_ancient.mli
@@ -1,0 +1,9 @@
+val is_available : bool
+
+(* Trimmed ocaml-ancient library signature *)
+
+type 'a ancient
+
+val mark : 'a -> 'a ancient
+val follow : 'a ancient -> 'a
+val delete : 'a ancient -> unit