--- a/lib/ancient/gw_ancient.wrapped.ml
+++ b/lib/ancient/gw_ancient.wrapped.ml
@@ -1,0 +1,3 @@
+let is_available = true
+
+include Ancient