--- a/lib/gwdb-legacy/dbdisk.mli
+++ b/lib/gwdb-legacy/dbdisk.mli
@@ -304,8 +304,6 @@ type 'a record_access = {
   get : int -> 'a;
   (* Same as [get] but doesn't consider pending patches *)
   get_nopending : int -> 'a;
-  (* Set the nth element of array *)
-  set : int -> 'a -> unit;
   (* Return length of an array that by default takes into account
      commited patches *)
   mutable len : int;