This PR proposes a new option for `gwd`: `-cache-in-memory <DATABASE>`
It allows to specify that a given database should be read and mapped in memory at launch time to avoid file accesses on a per-request basis. This proves to have a significant impact on response time for requests on huge databases.

## What was

Before this patch, `gwd` worked as follows:
The server launched with no need to be aware of the available databases. Each request handling would open the relevant database by opening some file descriptors and building some closures around them for data fetching functions used throughout the codebase. 
This means that each required piece of data would be read from disk, through systems calls and deserializations, both time-consuming operations.

Also, to handle requests, the server would fork its process for each requests, which is fine, as it ensure the main process would not stall or crash with dubious requests. However, memory generated in child processes is not available to the main, or other children.
In particular, this means that `gwd` won't take advantage of the internal caching mechanisms of the codebase, every loaded (and locally cached) piece of data will be dropped when the request process terminates, leaving every other requests to reload data from the disk.

## What is

The introduced option will make `gwd` open the given databases at launch and force-load (disk read and deserialization) all of its data in the already existing cache mechanism. Then, it uses the [ocaml-ancient](https://github.com/UnixJunkie/ocaml-ancient) library to move these arrays outside of the GC-managed heap as read-only data. This allows these data to be read by every subsequent child processes, without being duplicated at each fork.

This approach is possible thanks to the way databases are structured, the few files written by `gwd` (mainly the `patches` file) are unaltered by this PR, only the read-only data, which usually constitute the vast majority of the database.

## Some numbers

To illustrate the benefits of this PR, I tried to measure the response time of various queries to a server hosted on my poor man's PC. On a database with several million individuals registered, I compared the results of this branch, and of the commit its base on (this repo `master`, with a few commits behind).

Response time was measured by timing `curl` queries, several time to have an rough estimate on variability and to make sure that database files were at least cached by the kernel to discard this variability factor.

| Query | With caching | On `master` | Speed-up |
|--------|--------|--------|--------|
|`?m=N&tri=A`|15.9s|18.4s|1.16x|
|`?m=P&tri=A`|11.2s|16.2s|1.45x|
|`?m=PPS&bi=on&ba=on&ma=on&de=on&bu=on`|31.8s|164.8s|5.18x|
|`?m=LB&k=20`|1.6s|5.9s|3.69x|
|`?m=AN`|1.0s|5.9s|5.90x|
|`?m=NOTES`|4.3s|77.9s|18.12x|
|`?m=S&n=&p=albert`|0.6s|5.3s|8.83x|
|`?m=TT&t=&p=`|1.4s|11.0s|7.86x|
|`?m=POP_PYR`|1.1s|6.1s|5.55x|
|`?em=R&ei=5245870&long=on&et=S&p=albert&n=dupont`|0.2s|0.3s|1.50x|
|`?i=4722419`|0.3s|0.2s|0.67x|

Below the second, variability is too much for the comparison to be significant. This is also why I rounded the timing to the first decimal.

## The catch

All this comes with a few drawbacks: 
- `gwd` with caching is slower to launch because of all the preloading and moving memory around (with my (large) database, it takes approximately 30s for the server to be ready to accept requests).
- to properly use the memory created by `ocaml-ancient`, I needed to use a variant of the OCaml compiler with `no-naked-pointers`, hence the changes in `dune-project` and the opam file, and the reason CI will fail.
- `ocaml-ancient` itself is not available for Windows build.
- as of yet, `ocaml-ancient` is not compatible with OCaml 5

I understand that build accessibility might be an concern, and for that I'm thinking of a way to make this feature and its added dependencies optional at build time, as to not pollute the current distribution workflow.

## How to build

Amid the constraints above, this patch does not really alter the build workflow. All you need is to make sure to use an opam switch with the right compiler variant. 
`opam switch create . --deps-only` in this branch root directory should setup up the right environment to build everything as usual.

#

All those considerations aside, the code itself is in an reasonable state, and early reviews would be greatly appreciated.