@Keryan-dev 2024-04-03
It is not. If it stays on the heap, it will be duplicated by each fork. On a huge database, not only there is the memory cost issue but it also induces a lot of latency due to the copy, for each request.

Ancient is a way to make all those data shared by every child processes, and the only think duplicated by forks are the pointers to that part of memory.

There could be other options of course. For example, prefetching the arrays normally would work if the queries were handled with multi-threading. However, as I said, this involves an entire class of issues brought by concurrency, and error handling, because query-handling could no longer simply crash at failure.