@canonici 2024-04-03
Ok, I was under the impression that copy would not be forced to occur based on copy on write in the child process.