@Keryan-dev 2024-04-05
There. Proper conditional compilation that will add the feature to builds only in environment where its available. This also mean that opam will no longer install those dependencies with the command I mentioned, they need to be installed manually.

I also added an option to `configure.ml` : `--gwd-caching` which do nothing on itself but to emit warnings when the feature is not available in the current build environment. If it is available, `configure.ml` will set it up either way. The option is only useful when we want to make sure the feature will be built.

Feel free to nitpick on the code all you want before approving the merge.