@hgouraud 2024-04-22
Is this precaution really necessary? bases are located in the directory identified by the -bd parameter (./bases by default).
using Secure.bd_r or Secure.base_dir () should be sufficient. No??
Searching for bases across multiple directories is not offered as far as I know.