@Keryan-dev 2024-04-22
I just tried it and it is absolutely possible to input `dir/base` in the root page. The server will figure it out even if the base is not suggested.
And again, now that the user can input path themselves, the server should be wary of that.

But you a right, if the intended behavior is to consider only bases in the specified path without subdirs then its an issue of enforcing this constraint in the code, rather than a more general way of locating bases canonically.