@hgouraud 2024-04-22
would it be sufficient to say
   ` let bfname = Filename.concat (Secure.base_dir ()) bname in`